requirements.txt
geopandas
shapely
google-cloud-bigquery
folium
selenium
db_dtypes
